factored out logic from component
