+++ src/components/InstructionsAreaTextSelection.js
@@ -1,0 +1,31 @@
+import React from 'react';
+import * as windowSelectionHelpers from '../helpers/windowSelectionHelpers';
+const ELLIPSIS = '…';
+
+export default ({ selections, verseText }) => {
+  if (windowSelectionHelpers.shouldRenderEllipsis(selections, verseText)) {
+    return (
+      <QuoatationMarks>
+        {selections[0].text.trim()}
+        <Ellipsis />
+        {selections[selections.length - 1].text.trim()}
+      </QuoatationMarks>
+    );
+  } else {
+    return (
+      <QuoatationMarks>
+        {selections.map((selection, index) =>
+          <span key={index}>
+            <strong style={{ color: 'var(--accent-color)' }}>
+              {`${selection.text.trim()}`}
+            </strong>
+            {selections[index + 1] ? <span>{" "}</span> : null}
+          </span>
+        )}
+      </QuoatationMarks>
+    );
+  }
+};
+
+export const QuoatationMarks = ({ children }) => <strong style={{ color: 'var(--accent-color)' }}>{'"'}{children}{'"'}</strong>;
+export const Ellipsis = () => <strong style={{ color: 'var(--accent-color)' }}>{` ${ELLIPSIS} `}</strong>;
