+++ src/helpers/windowSelectionHelpers.js
@@ -112,3 +112,28 @@ export const getPrescedingTextFromElementSiblings = (element, windowSelection) =
   }
   return prescedingText;
 };
+
+/**
+ * This is a helper method to determine if the selection needs an ellipsis in
+ * between the selected words or not.
+ * @param {Array} selections - Array of word objects that the user selected
+ * @param {string} verseText - The entire verse string from the current check
+ * @returns {boolean} - Whether or not the View should display an ellipsis
+ */
+export function shouldRenderEllipsis(selections, verseText) {
+  /** Need to get the the words and occurrence of the selected edge words */
+  const endSelectedWord = selections[selections.length - 1].text.trim();
+  const endSelectedWordOccurrence = selections[selections.length - 1].occurrence;
+  const beginningSelectedWord = selections[0].text.trim();
+  const beginningSelectedWordOccurrence = selections[0].occurrence;
+
+  /** Using the occurrences to get the actual index of the word vs
+   *  the first time it appears in verse text */
+  const indexOfBeginningSelection = verseText.split(beginningSelectedWord, beginningSelectedWordOccurrence).join(beginningSelectedWord).length;
+  const indexOfEndSelection = verseText.split(endSelectedWord, endSelectedWordOccurrence).join(endSelectedWord).length;
+
+  /** Checking the text in between selected words for a non space character */
+  const textBetweenSelection = verseText.substring(indexOfBeginningSelection + beginningSelectedWord.length, indexOfEndSelection);
+  /** If the end index is the same as the beginning then it is the first word */
+  return (indexOfEndSelection !== indexOfBeginningSelection) && textBetweenSelection.match(/\S/);
+}
